pyproject.toml
[project]
name = "paroquant"
version = "0.1.0"
requires-python = ">=3.11"
description = "ParoQuant — Pairwise Rotation Quantization for LLMs"
license = { text = "MIT" }
dependencies = [
    "rich",
]

[project.optional-dependencies]
transformers = [
    "torch>=2.8",
    "torchvision",
    "transformers>=4.55",
    "autoawq",
]
vllm = [
    "vllm>=0.15",
]
mlx = [
    "mlx",
    "mlx-lm",
    "mlx-vlm",
    "numpy",
    "transformers",
    "safetensors",
    "huggingface-hub",
    "sentencepiece",
    "protobuf",
]
optim = [
    "paroquant[transformers]",
    "datasets",
    "simple_parsing",
    "tqdm",
]
eval = [
    "lm_eval",
    "zstandard",
]
dev = [
    "pytest",
    "pre-commit",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["paroquant*"]

[tool.black]
line-length = 120
target-version = ["py311"]
